Bunch of bugs fixed
grunt was not hitting
abomination was killing instantly
fight mode text errors
flight was not showing mana
state transition camera zoom problem
holly ligt health border problems
score and health text was overlapping
score and wave text was overlapping
abomination texture problem

+++ src/core/src/com/mygdx/game/Score.java
@@ -25,6 +25,6 @@ public class Score implements Comparable<Score>, Serializable{
 
     @Override
     public String toString() {
-        return playerName + " " + score;
+        return playerName + " " + (int) score;
     }
 }
